backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
firebase-admin>=6.5.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
yt-dlp>=2025.1.0
openai>=1.0.0
google-cloud-texttospeech>=2.0.0
python-multipart>=0.0.9
google-api-python-client>=2.0.0
